+++ scratch.js
@@ -7,9 +7,10 @@ const spacetime = require('/Users/spencer/mountain/spacetime')
 const fmt = iso => (iso ? spacetime(iso).format('{day-short} {nice} {year}') : '-')
 nlp.verbose('date')
 
-let doc = nlp('3:30pm-3:45').debug()
+// let doc = nlp('tuesdays').debug()
+let doc = nlp('any tuesday').debug()
 let dates = doc.dates().get()
-// console.log(doc.dates().get(0))
+console.log(dates[0])
 dates.forEach(date => {
   console.log('start: ', fmt(date.start))
   console.log('  end: ', fmt(date.end))